Add logic that creates a grid based on user
Created a button that prompts the user for new
side length of each square within the grid.

Reworked sizing of container and original squares.

Created logic that changes the sizing of the
squares within the grids based on the user input
and updates it each time a new grid is created.

Created logic to create entire new grid using
the same logic as the original grid.

Updated styles.